requirements.txt
PyGithub==2.3.0
discord.py==1.7.3
requests==2.31.0
